refactor(test): 인수테스트 내 restAssured 사용 부분 따로 클래스 빼서 관리하도록 리팩토링
